perf/requirements.txt
apiritif==0.8.1
Appium-Python-Client==0.40
astunparse==1.6.2
atomicwrites==1.3.0
attrs==19.1.0
bzt==1.13.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
cssselect==1.0.3
decorator==4.4.0
EasyProcess==0.2.5
Flask==1.0.2
fuzzyset==0.0.19
gevent==1.4.0
greenlet==0.4.15
hdrpy==0.3.3
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpath-rw==1.4.0
locustio==0.11.0
lxml==4.3.3
MarkupSafe==1.1.1
more-itertools==7.0.0
msgpack==0.6.1
nose==1.3.7
numpy==1.16.3
pluggy==0.9.0
ply==3.11
progressbar33==2.4
psutil==5.6.1
py==1.8.0
pytest==4.4.1
python-Levenshtein==0.12.0
PyVirtualDisplay==0.2.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
terminaltables==3.1.0
texttable==1.6.1
unicodecsv==0.14.1
urllib3==1.24.2
urwid==2.0.1
Werkzeug==0.15.2
